I did a lot of stuff without committing. This one is a monster. I didn't commit because I didn't reach a comfortable checkpoint. There's still a lot being done. But, here's what I can recall having done at the time of this commit.
- Added JSDoc
- Changed favicon to Wall St $
- Worked on Certifications

The Certification work and documentation is the star of the show, this commit. I've been in a cycle of inspiration, conception, creation, and documentation. The code is both leading and following the inspirational and creative process. So it's all a big mess right now. Nevertheless, there's a lot of work here I'd like to preserve in a safe place... hence this commit.

Sorry.